+++ gitlab/CHANGELOG.md
@@ -1,4 +1,7 @@
 # GitLab Connector Change Log
 
+## 2022/08/04
+* Improved error handling and logging
+
 ## 2022/4/18
 * OS environment variable `COOKIE_API_TOKEN` changed to `VEZA_API_KEY`